TestFW now prints state of all sensors on proto box
The proto box design as-is has a joystick, buttons, a linearPotentiometer, and a ultrasonic rangefinder- this FW has been used to ensure everything is working on the boxes before they are handed off for feature development.

All of this is done via printing to serial- this FW is not very useful if DEBUG is not defined.