Remove deprecated projects
No reason to keep these Arduino projects, given that we have this in Git.  These are just going to be confusing and distracting.

I do think there would be value in a diagnostic program that prints out the value of everything, but that needs to be its own polished project.