Commit attempt at periodic restart
This code should restart the teensy at some periodic intervals based on user preference, and should allow for user to disable automatic reboot as well, if desired.

User can only select one value, and failure to select a value should result in compilation errors.

Finally, the REBOOT_MIN_IN_MSEC is not super realistic, but is intended for TEST ONLY!