Simple case of NeoStick displaying ultrasonic value for percussion/melodic stations
I'm going to make another separate Pull Request for a more complex version, one that will use the StationType in EEPROM to set the strip colors to match the station arcade button colors. But for now this will work in a pinch and is so simple I have high expectations that it will work off the bat.

NOTE: For NeoPixel.cpp, I have to flip the value to make sense now for ultrasonic range as opposed to countdown time. Could re-write the function but this is fine enough.. Maybe the complex case will redo the array order, since it already has to redo the array contents