PIR Presence option 1: Remove presence MIDI CC messages of any form.
So with this branch, the PIR message stuff will still cause the LEDs to ramp up and ramp down, but will no longer send a MIDI message- so will not be controlling volume of any subsystems or anything.

This is 1 option (of 2) for how to deal with this PIR presence issue. There will be another PR with another option, but we should select one of these branches and just delete the other.